+++ domain/interface/action-row.builder.interface.ts
