+++ domain/interface/interaction.interface.ts
